+++ braintree/__init__.py
@@ -24,6 +24,7 @@ from braintree.disbursement import Disbursement
 from braintree.discount import Discount
 from braintree.discount_gateway import DiscountGateway
 from braintree.dispute import Dispute
+from braintree.dispute_search import DisputeSearch
 from braintree.environment import Environment
 from braintree.error_codes import ErrorCodes
 from braintree.error_result import ErrorResult