+++ braintree/dispute_gateway.py
@@ -5,6 +5,9 @@ from braintree.dispute_details import DisputeEvidence
 from braintree.error_result import ErrorResult
 from braintree.successful_result import SuccessfulResult
 from braintree.exceptions.not_found_error import NotFoundError
+from braintree.paginated_result import PaginatedResult
+from braintree.paginated_collection import PaginatedCollection
+from braintree.resource_collection import ResourceCollection
 
 class DisputeGateway(object):
     def __init__(self, gateway):
@@ -84,3 +87,25 @@ class DisputeGateway(object):
                 return SuccessfulResult()
         except NotFoundError:
             raise NotFoundError("evidence with id " + repr(evidence_id) + " for dispute with id " + repr(dispute_id) + " not found")
+
+    def search(self, *query):
+        self.search_criteria = self.__criteria(query)
+
+        pc = PaginatedCollection(self.__fetch_disputes)
+        return SuccessfulResult({"disputes": pc})
+
+    def __fetch_disputes(self, page):
+        response = self.config.http().post(self.config.base_merchant_path() + "/disputes/advanced_search?page=" + str(page), {"search": self.search_criteria})
+        body = response["disputes"]
+
+        disputes = [Dispute(item) for item in ResourceCollection._extract_as_array(response["disputes"], "dispute")]
+        return PaginatedResult(body["total_items"], body["page_size"], disputes)
+
+    def __criteria(self, query):
+        criteria = {}
+        for term in query:
+            if criteria.get(term.name):
+                criteria[term.name] = dict(list(criteria[term.name].items()) + list(term.to_param().items()))
+            else:
+                criteria[term.name] = term.to_param()
+        return criteria