+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add braintree.Dispute.finalize
 * Add braintree.Dispute.find
 * Add braintree.Dispute.remove_evidence
+* Add braintree.Dispute.search
 
 ## 3.37.2
 * Fix a bug where a null value for `amount` in `CreditCardVerification` would result in a `ValueError`